requirements.txt
pdfminer.six
numpy
scikit-learn